Resolução desafio 35 - Int para Float

+++ Exercícios/Desafios/desafio35.py
@@ -1,6 +1,6 @@
-r1 = int(input('Digite a primeira reta: '))
-r2 = int(input('Digite a segunda reta: '))
-r3 = int(input('Digite a terceira reta: '))
+r1 = float(input('Digite a primeira reta: '))
+r2 = float(input('Digite a segunda reta: '))
+r3 = float(input('Digite a terceira reta: '))
 
 possivel = r3 > r1 - r2 and r3 < r1 + r2
